+++ decidim-core/app/permissions/decidim/permissions.rb
@@ -18,6 +18,7 @@ module Decidim
       follow_action?
       notification_action?
       conversation_action?
+      user_group_action?
 
       permission_action
     end
@@ -89,6 +90,11 @@ module Decidim
 
       conversation = context.fetch(:conversation, nil)
       toggle_allow(conversation.participants.include?(user))
+    end
+
+    def user_group_action?
+      return unless permission_action.subject == :user_group
+      return allow! if [:create].include?(permission_action.action)
     end
 
     def user_can_admin_component?